backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voice-interviewer-backend"
version = "0.1.0"
description = "Minimal Pipecat backend for a local real-time interviewer demo"
requires-python = ">=3.12,<3.13"
dependencies = [
  "pipecat-ai[runner,webrtc,deepgram,anthropic,elevenlabs,silero]>=0.0.108,<0.0.109",
  "python-dotenv>=1.0.1,<2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
